[Refactor/#116]: TabMenu 분리 및 상점 페이지에 적용
